OAuthフローでのインストール、アプリのイベントでの動作確認済み #21
